+++ lib/builtin/pql/pql.dart
@@ -2,4 +2,4 @@ export 'src/backlinks_controller.dart';
 export 'src/backlinks_view.dart';
 export 'src/extension.dart';
 export 'src/pql_controller.dart';
-export 'src/pql_panel_view.dart';
+export 'src/pql_search_body.dart';